# sdk/log: Assign fltrProcessors on provider creation instead of lazy